doc-classify/requirements.txt
langchain
langchain-google-genai
PyMuPDF
Pillow
python-dotenv
google-generativeai
aiofiles
fastapi
python-multipart
uvicorn
langsmith